// src/lib/api/models/PromoCodePublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PromoCodePublic
 */
export interface PromoCodePublic {
    /**
     * 
     * @type {string}
     * @memberof PromoCodePublic
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PromoCodePublic
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof PromoCodePublic
     */
    smsCost: string;
    /**
     * 
     * @type {boolean}
     * @memberof PromoCodePublic
     */
    isActive: boolean;
    /**
     * 
     * @type {Date}
     * @memberof PromoCodePublic
     */
    expiresAt: Date | null;
    /**
     * 
     * @type {number}
     * @memberof PromoCodePublic
     */
    maxUses: number | null;
    /**
     * 
     * @type {number}
     * @memberof PromoCodePublic
     */
    currentUses: number;
    /**
     * 
     * @type {string}
     * @memberof PromoCodePublic
     */
    description: string | null;
    /**
     * 
     * @type {Date}
     * @memberof PromoCodePublic
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the PromoCodePublic interface.
 */
export function instanceOfPromoCodePublic(value: object): value is PromoCodePublic {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('smsCost' in value) || value['smsCost'] === undefined) return false;
    if (!('isActive' in value) || value['isActive'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('maxUses' in value) || value['maxUses'] === undefined) return false;
    if (!('currentUses' in value) || value['currentUses'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function PromoCodePublicFromJSON(json: any): PromoCodePublic {
    return PromoCodePublicFromJSONTyped(json, false);
}

export function PromoCodePublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): PromoCodePublic {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'code': json['code'],
        'smsCost': json['sms_cost'],
        'isActive': json['is_active'],
        'expiresAt': (json['expires_at'] == null ? null : new Date(json['expires_at'])),
        'maxUses': json['max_uses'],
        'currentUses': json['current_uses'],
        'description': json['description'],
        'createdAt': (new Date(json['created_at'])),
    };
}

export function PromoCodePublicToJSON(json: any): PromoCodePublic {
    return PromoCodePublicToJSONTyped(json, false);
}

export function PromoCodePublicToJSONTyped(value?: PromoCodePublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'code': value['code'],
        'sms_cost': value['smsCost'],
        'is_active': value['isActive'],
        'expires_at': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'max_uses': value['maxUses'],
        'current_uses': value['currentUses'],
        'description': value['description'],
        'created_at': value['createdAt'].toISOString(),
    };
}