// src/lib/api/models/PromoCodeCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PromoCodeCreate
 */
export interface PromoCodeCreate {
    /**
     * Unique promo code
     * @type {string}
     * @memberof PromoCodeCreate
     */
    code: string;
    /**
     * Special SMS cost (e.g., '25', '18')
     * @type {string}
     * @memberof PromoCodeCreate
     */
    smsCost: string;
    /**
     * 
     * @type {boolean}
     * @memberof PromoCodeCreate
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof PromoCodeCreate
     */
    expiresAt?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof PromoCodeCreate
     */
    maxUses?: number | null;
    /**
     * 
     * @type {string}
     * @memberof PromoCodeCreate
     */
    description?: string | null;
}

/**
 * Check if a given object implements the PromoCodeCreate interface.
 */
export function instanceOfPromoCodeCreate(value: object): value is PromoCodeCreate {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('smsCost' in value) || value['smsCost'] === undefined) return false;
    return true;
}

export function PromoCodeCreateFromJSON(json: any): PromoCodeCreate {
    return PromoCodeCreateFromJSONTyped(json, false);
}

export function PromoCodeCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): PromoCodeCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'],
        'smsCost': json['sms_cost'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'expiresAt': json['expires_at'] == null ? undefined : (new Date(json['expires_at'])),
        'maxUses': json['max_uses'] == null ? undefined : json['max_uses'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function PromoCodeCreateToJSON(json: any): PromoCodeCreate {
    return PromoCodeCreateToJSONTyped(json, false);
}

export function PromoCodeCreateToJSONTyped(value?: PromoCodeCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'sms_cost': value['smsCost'],
        'is_active': value['isActive'],
        'expires_at': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'max_uses': value['maxUses'],
        'description': value['description'],
    };
}