// src/lib/api/models/PromoCodeUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PromoCodeUpdate
 */
export interface PromoCodeUpdate {
    /**
     * 
     * @type {string}
     * @memberof PromoCodeUpdate
     */
    smsCost?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof PromoCodeUpdate
     */
    isActive?: boolean | null;
    /**
     * 
     * @type {Date}
     * @memberof PromoCodeUpdate
     */
    expiresAt?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof PromoCodeUpdate
     */
    maxUses?: number | null;
    /**
     * 
     * @type {string}
     * @memberof PromoCodeUpdate
     */
    description?: string | null;
}

/**
 * Check if a given object implements the PromoCodeUpdate interface.
 */
export function instanceOfPromoCodeUpdate(value: object): value is PromoCodeUpdate {
    return true;
}

export function PromoCodeUpdateFromJSON(json: any): PromoCodeUpdate {
    return PromoCodeUpdateFromJSONTyped(json, false);
}

export function PromoCodeUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): PromoCodeUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'smsCost': json['sms_cost'] == null ? undefined : json['sms_cost'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'expiresAt': json['expires_at'] == null ? undefined : (new Date(json['expires_at'])),
        'maxUses': json['max_uses'] == null ? undefined : json['max_uses'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function PromoCodeUpdateToJSON(json: any): PromoCodeUpdate {
    return PromoCodeUpdateToJSONTyped(json, false);
}

export function PromoCodeUpdateToJSONTyped(value?: PromoCodeUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'sms_cost': value['smsCost'],
        'is_active': value['isActive'],
        'expires_at': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'max_uses': value['maxUses'],
        'description': value['description'],
    };
}